Move `ComponentRules` and related logic into a top-level plugin
Replicated archetypes and removals buffer update systems coupled with
compoent rules. So I think they belong into their own plugin.

Since `removal_buffer` module now contains only a single struct, I
combined it with `despawn_buffer` module and called it `world_buffers`.

Tests for removals now integrational since it's a part of public API.
