+++ src/component_rules.rs
@@ -5,6 +5,8 @@ use bevy::{
         archetype::{ArchetypeGeneration, Archetypes},
         component::ComponentId,
         entity::MapEntities,
+        event::ManualEventReader,
+        removal_detection::{RemovedComponentEntity, RemovedComponentEvents},
     },
     prelude::*,
     ptr::Ptr,
@@ -13,13 +15,97 @@ use bevy::{
 use bincode::{DefaultOptions, Options};
 use serde::{de::DeserializeOwned, Serialize};
 
-use super::{
-    replication_fns::{
-        DeserializeFn, RemoveFn, RemoveFnId, ReplicationFns, SerdeFns, SerdeFnsId, SerializeFn,
+use crate::{
+    client::client_mapper::{ClientMapper, ServerEntityMap},
+    core::{
+        replicated_archetypes::{ReplicatedArchetype, ReplicatedArchetypes, ReplicatedComponent},
+        replication_fns::{
+            DeserializeFn, RemoveFn, RemoveFnId, ReplicationFns, SerdeFns, SerdeFnsId, SerializeFn,
+        },
+        replicon_tick::RepliconTick,
     },
-    replicon_tick::RepliconTick,
+    server::world_buffers::{DespawnBuffer, RemovalBuffer},
+    server_running, ServerSet,
 };
-use crate::client::client_mapper::{ClientMapper, ServerEntityMap};
+
+pub struct ComponentRulesPlugin;
+
+impl Plugin for ComponentRulesPlugin {
+    fn build(&self, app: &mut App) {
+        app.init_resource::<ComponentRules>().add_systems(
+            PostUpdate,
+            (
+                Self::update_replicated_archetypes.in_set(ServerSet::UpdateArchetypes),
+                Self::buffer_removals
+                    .in_set(ServerSet::BufferRemovals)
+                    .run_if(server_running),
+            ),
+        );
+    }
+}
+
+impl ComponentRulesPlugin {
+    fn update_replicated_archetypes(
+        archetypes: &Archetypes,
+        mut replicated_archetypes: ResMut<ReplicatedArchetypes>,
+        mut component_rules: ResMut<ComponentRules>,
+    ) {
+        let old_generation = component_rules.update_generation(archetypes);
+
+        // Archetypes are never removed, iterate over newly added since the last update.
+        let marker_id = replicated_archetypes.marker_id();
+        for archetype in archetypes[old_generation..]
+            .iter()
+            .filter(|archetype| archetype.contains(marker_id))
+        {
+            let mut replicated_archetype = ReplicatedArchetype::new(archetype.id());
+            for component_id in archetype.components() {
+                let Some(&serde_id) = component_rules.serde_ids().get(&component_id) else {
+                    continue;
+                };
+
+                // SAFETY: component ID obtained from this archetype.
+                let storage_type =
+                    unsafe { archetype.get_storage_type(component_id).unwrap_unchecked() };
+
+                let replicated_component = ReplicatedComponent {
+                    component_id,
+                    storage_type,
+                    serde_id,
+                };
+
+                // SAFETY: Component ID and storage type obtained from this archetype,
+                // serde functions ID points to existing functions from `ComponentRules`.
+                unsafe { replicated_archetype.add_component(replicated_component) };
+            }
+
+            // SAFETY: Archetype ID corresponds to a valid archetype.
+            unsafe { replicated_archetypes.add_archetype(replicated_archetype) };
+        }
+    }
+
+    fn buffer_removals(
+        mut readers: Local<HashMap<ComponentId, ManualEventReader<RemovedComponentEntity>>>,
+        remove_events: &RemovedComponentEvents,
+        mut removal_buffer: ResMut<RemovalBuffer>,
+        component_rules: Res<ComponentRules>,
+        despawn_buffer: Res<DespawnBuffer>,
+    ) {
+        for (&component_id, &serde_id) in component_rules.remove_ids() {
+            for removals in remove_events.get(component_id).into_iter() {
+                let reader = readers.entry(component_id).or_default();
+                for entity in reader
+                    .read(removals)
+                    .cloned()
+                    .map(Into::into)
+                    .filter(|entity| !despawn_buffer.contains(entity))
+                {
+                    removal_buffer.insert(entity, serde_id);
+                }
+            }
+        }
+    }
+}
 
 pub trait AppReplicationExt {
     /// Marks component for replication.
