+++ src/core.rs
@@ -1,5 +1,4 @@
 pub mod common_conditions;
-pub mod component_rules;
 pub mod replicated_archetypes;
 pub mod replication_fns;
 pub mod replicon_channels;
@@ -7,7 +6,6 @@ pub mod replicon_tick;
 
 use bevy::prelude::*;
 
-use component_rules::ComponentRules;
 use replicated_archetypes::ReplicatedArchetypes;
 use replication_fns::ReplicationFns;
 use replicon_channels::RepliconChannels;
@@ -22,8 +20,7 @@ impl Plugin for RepliconCorePlugin {
             .init_resource::<RepliconTick>()
             .init_resource::<ReplicatedArchetypes>()
             .init_resource::<RepliconChannels>()
-            .init_resource::<ReplicationFns>()
-            .init_resource::<ComponentRules>();
+            .init_resource::<ReplicationFns>();
     }
 }
