+++ src/core/replicated_archetypes.rs
@@ -16,7 +16,7 @@ use crate::core::{replication_fns::SerdeFnsId, Replication};
 /// - Register 'serde' and 'remove' functions inside [`ReplicationFns`](crate::core::replication_fns::ReplicationFns).
 /// - Update this struct for all newly added archetypes in
 /// [`ServerSet::UpdateArchetypes`](crate::server::ServerSet::UpdateArchetypes) using the registered function IDs.
-/// - Update [`RemovalBuffer`](crate::server::removal_buffer::RemovalBuffer) in
+/// - Update [`RemovalBuffer`](crate::server::world_buffers::RemovalBuffer) in
 /// [`ServerSet::BufferRemovals`](crate::server::ServerSet::BufferRemovals) when the rule components should be removed.
 #[derive(Resource)]
 pub struct ReplicatedArchetypes {
