--- src/server/despawn_buffer.rs
@@ -1,61 +1,0 @@
-use bevy::prelude::*;
-
-use super::ServerSet;
-use crate::core::{common_conditions::server_running, Replication};
-
-/// Treats removals of [`Replication`] component as despawns and stores them into [`DespawnBuffer`] resource.
-///
-/// Used to avoid missing events in case the server's tick policy is not [`TickPolicy::EveryFrame`].
-pub(super) struct DespawnBufferPlugin;
-
-impl Plugin for DespawnBufferPlugin {
-    fn build(&self, app: &mut App) {
-        app.init_resource::<DespawnBuffer>().add_systems(
-            PostUpdate,
-            Self::buffer_despawns
-                .in_set(ServerSet::BufferDespawns)
-                .run_if(server_running),
-        );
-    }
-}
-
-impl DespawnBufferPlugin {
-    fn buffer_despawns(
-        mut removed_replications: RemovedComponents<Replication>,
-        mut despawn_buffer: ResMut<DespawnBuffer>,
-    ) {
-        for entity in removed_replications.read() {
-            despawn_buffer.push(entity);
-        }
-    }
-}
-
-/// Buffer with all despawned entities.
-///
-/// Should be cleaned up manually.
-#[derive(Default, Resource, Deref, DerefMut)]
-pub(crate) struct DespawnBuffer(Vec<Entity>);
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::server::replicon_server::RepliconServer;
-
-    #[test]
-    fn despawns() {
-        let mut app = App::new();
-        app.add_plugins(DespawnBufferPlugin)
-            .init_resource::<RepliconServer>();
-
-        app.world.resource_mut::<RepliconServer>().set_running(true);
-
-        app.update();
-
-        app.world.spawn(Replication).despawn();
-
-        app.update();
-
-        let despawn_buffer = app.world.resource::<DespawnBuffer>();
-        assert_eq!(despawn_buffer.len(), 1);
-    }
-}
